# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hydrus-go HTTP API",
    "version": "0.1.0",
    "description": "Machine-readable description of the currently implemented hydrus-go daemon HTTP surface. This spec intentionally describes the implemented migration slice, not full Hydrus Client API parity. Protected operations prefer header credentials; the current implementation also accepts the same access/session key names as query parameters for compatibility."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:45869",
      "description": "Default local hydrus-go daemon"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "HydrusAccessKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Hydrus-Client-API-Access-Key",
        "description": "64-character hex access key. Preferred auth mechanism for protected endpoints."
      },
      "HydrusSessionKey": {
        "type": "apiKey",
        "in": "header",
        "name": "Hydrus-Client-API-Session-Key",
        "description": "64-character hex session key minted from /session_key. Preferred auth mechanism for protected endpoints."
      }
    },
    "schemas": {
...
      "VerifyAccessKeyResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ResponseMeta" },
          {
            "type": "object",
            "properties": {
              "name": { "type": "string" },
              "permits_everything": { "type": "boolean" },
              "basic_permissions": {
                "type": "array",
                "items": { "type": "integer" }
              },
              "human_description": { "type": "string" }
            },
            "required": ["name", "permits_everything", "basic_permissions", "human_description"]
          }
        ]
      },
      "SessionKeyResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ResponseMeta" },
          {
            "type": "object",
            "properties": {
              "session_key": {
                "type": "string",
                "pattern": "^[a-f0-9]{64}$"
              }
            },
            "required": ["session_key"]
          }
        ]
      },
...
      "ServicesResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/ResponseMeta" },
          {
            "type": "object",
            "properties": {
              "services": {
                "type": "object",
                "additionalProperties": { "$ref": "#/components/schemas/LegacyService" }
              },
              "services_v2": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/Service" }
              }
            },
            "required": ["services", "services_v2"],
            "additionalProperties": true
          }
        ]
      },
...
      "TrashRequest": {
        "type": "object",
        "properties": {
          "file_id": { "type": "integer", "format": "int64" }
        },
        "required": ["file_id"],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/verify_access_key": {
      "get": {
        "tags": ["auth"],
        "summary": "Verify access/session key",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "responses": {
          "200": {
            "description": "Authenticated principal details",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VerifyAccessKeyResponse" }
              }
            }
          },
          "401": { "description": "Missing credentials", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "403": { "description": "Invalid credentials or permissions", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "419": { "description": "Expired session key", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/session_key": {
      "get": {
        "tags": ["auth"],
        "summary": "Mint session key",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "responses": {
          "200": {
            "description": "Session key created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionKeyResponse" }
              }
            }
          },
          "401": { "description": "Missing credentials", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "403": { "description": "Invalid access key", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "419": { "description": "Expired session key", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/get_services": {
      "get": {
        "tags": ["services"],
        "summary": "List services",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "responses": {
          "200": { "description": "Discovery-visible services", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ServicesResponse" } } } },
          "401": { "description": "Missing credentials", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "403": { "description": "Insufficient permissions", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "419": { "description": "Expired session", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "500": { "description": "Service list unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/get_service": {
      "get": {
        "tags": ["services"],
        "summary": "Lookup one service",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "parameters": [
          { "name": "service_key", "in": "query", "schema": { "type": "string" }, "description": "Hex service key. Provide either service_key or service_name." },
          { "name": "service_name", "in": "query", "schema": { "type": "string" }, "description": "Public discovery-visible service name. Provide either service_key or service_name." }
        ],
        "responses": {
          "200": { "description": "Service lookup result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ServiceLookupResponse" } } } },
          "400": { "description": "Invalid query or undiscoverable service", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "404": { "description": "Service not found", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/get_files/file_metadata": {
      "get": {
        "tags": ["metadata"],
        "summary": "Hydrus-compatible file metadata",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "parameters": [
          { "name": "hash", "in": "query", "schema": { "type": "string" }, "description": "One SHA-256 hex hash." },
          { "name": "hashes", "in": "query", "schema": { "type": "string" }, "description": "JSON-encoded array of SHA-256 hex hashes." },
          { "name": "file_id", "in": "query", "schema": { "type": "integer", "format": "int64" }, "description": "One file id." },
          { "name": "file_ids", "in": "query", "schema": { "type": "string" }, "description": "JSON-encoded array of file ids." },
          { "name": "only_return_identifiers", "in": "query", "schema": { "type": "boolean" } },
          { "name": "only_return_basic_information", "in": "query", "schema": { "type": "boolean" } },
          { "name": "include_services_object", "in": "query", "schema": { "type": "boolean", "default": true } },
          { "name": "hide_service_keys_tags", "in": "query", "schema": { "type": "boolean", "default": true } },
          { "name": "include_blurhash", "in": "query", "schema": { "type": "boolean" } },
          { "name": "include_milliseconds", "in": "query", "schema": { "type": "boolean" } },
          { "name": "detailed_url_information", "in": "query", "schema": { "type": "boolean" } },
          { "name": "include_notes", "in": "query", "schema": { "type": "boolean" } },
          { "name": "create_new_file_ids", "in": "query", "schema": { "type": "boolean" } }
        ],
        "responses": {
          "200": { "description": "Metadata rows", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FileMetadataResponse" } } } },
          "400": { "description": "Invalid metadata request", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "404": { "description": "Requested file not found", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "Requested metadata mode not implemented", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/tags/autocomplete": {
      "get": {
        "tags": ["tags"],
        "summary": "Autocomplete tags",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "parameters": [
          { "name": "q", "in": "query", "required": false, "schema": { "type": "string" }, "description": "Tag prefix. Blank or omitted values return an empty suggestion list." },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 20, "minimum": 1 } }
        ],
        "responses": {
          "200": { "description": "Tag suggestions", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TagAutocompleteResponse" } } } },
          "400": { "description": "Invalid limit", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "500": { "description": "Autocomplete unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/library/recent": {
      "get": {
        "tags": ["browse"],
        "summary": "Recent local files",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "parameters": [
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100, "minimum": 1, "maximum": 250 } }
        ],
        "responses": {
          "200": { "description": "Recent browse page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RecentFilesResponse" } } } },
          "400": { "description": "Invalid offset or limit", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "Recent browse unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/library/search": {
      "get": {
        "tags": ["browse"],
        "summary": "Daemon-backed local-library search",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "parameters": [
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 100, "minimum": 1, "maximum": 250 } },
          { "name": "tags", "in": "query", "schema": { "type": "array", "items": { "type": "string" } }, "style": "form", "explode": true, "description": "Repeat this parameter for AND-tag queries." },
          { "name": "sort_by", "in": "query", "schema": { "type": "string", "enum": ["", "import_newest", "import_oldest", "size_desc", "size_asc"] } },
          { "name": "system_predicates[]", "in": "query", "schema": { "type": "array", "items": { "type": "string" } }, "style": "form", "explode": true, "description": "Repeat this parameter for supported size/width/height/favorite/resolution predicates." }
        ],
        "responses": {
          "200": { "description": "Search page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SearchFilesResponse" } } } },
          "400": { "description": "Invalid search request", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "Unsupported search mode", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/files/content": {
      "get": {
        "tags": ["assets"],
        "summary": "Stream managed original file",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "parameters": [
          { "name": "file_id", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "Managed original file stream", "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } } },
          "400": { "description": "Invalid file id", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "404": { "description": "Asset not found", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "Asset serving unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/files/thumbnail": {
      "get": {
        "tags": ["assets"],
        "summary": "Stream managed thumbnail",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "parameters": [
          { "name": "file_id", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64" } }
        ],
        "responses": {
          "200": { "description": "Managed thumbnail stream", "content": { "application/octet-stream": { "schema": { "type": "string", "format": "binary" } } } },
          "400": { "description": "Invalid file id", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "404": { "description": "Asset not found", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "Asset serving unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/files/trash": {
      "post": {
        "tags": ["trash"],
        "summary": "Trash one file",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TrashRequest" } } }
        },
        "responses": {
          "200": { "description": "File trashed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TrashResponse" } } } },
          "400": { "description": "Invalid trash request", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "404": { "description": "File not found", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "Trash unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/import/local_file": {
      "post": {
        "tags": ["imports"],
        "summary": "Import one daemon-local path",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportLocalFileRequest" } } }
        },
        "responses": {
          "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportResponse" } } } },
          "400": { "description": "Invalid import request", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "404": { "description": "Local path not found", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "Import unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/import/url": {
      "post": {
        "tags": ["imports"],
        "summary": "Import one direct file URL",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportURLRequest" } } }
        },
        "responses": {
          "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportResponse" } } } },
          "400": { "description": "Invalid URL import request or remote URL fetch failure", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "URL import unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/v1/import/upload": {
      "post": {
        "tags": ["imports"],
        "summary": "Import one uploaded file",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": { "type": "string", "format": "binary" },
                  "local_file_service_key": { "type": "string" },
                  "file_modified_at_ms": { "type": "string", "description": "Unix milliseconds as a string field." }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Import result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ImportResponse" } } } },
          "400": { "description": "Invalid upload form", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "413": { "description": "Upload exceeds body limit", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "501": { "description": "Upload import unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/add_tags/add_tags": {
      "post": {
        "tags": ["tags", "ptr"],
        "summary": "Stage PTR pending-add mappings",
        "description": "Current narrow implementation: exactly one service and only CONTENT_UPDATE_PEND (2) are supported.",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddTagsRequest" } } }
        },
        "responses": {
          "200": { "description": "Pending mappings staged", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AddTagsResponse" } } } },
          "400": { "description": "Invalid add-tags request", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "503": { "description": "PTR staging unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/manage_services/pending_counts": {
      "get": {
        "tags": ["ptr"],
        "summary": "Get locally staged PTR pending count",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "parameters": [
          { "name": "service_key", "in": "query", "schema": { "type": "string" }, "description": "Optional PTR service key. Defaults to the daemon-owned public tag repository service." }
        ],
        "responses": {
          "200": { "description": "Pending mapping count", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PendingCountResponse" } } } },
          "404": { "description": "PTR service not found", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "503": { "description": "PTR pending counts unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/manage_services/commit_pending": {
      "post": {
        "tags": ["ptr"],
        "summary": "Commit staged PTR pending mappings",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "requestBody": {
          "required": false,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommitPendingRequest" } } }
        },
        "responses": {
          "200": { "description": "Pending mappings committed", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommitPendingResponse" } } } },
          "400": { "description": "Invalid commit request", "content": { "text/plain": { "schema": { "type": "string" } } } },
          "503": { "description": "PTR pending commit unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/service/ptr/status": {
      "get": {
        "tags": ["ptr"],
        "summary": "Get daemon-owned PTR status",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "responses": {
          "200": { "description": "PTR status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PTRStatusResponse" } } } },
          "501": { "description": "PTR status unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/service/ptr/sync": {
      "post": {
        "tags": ["ptr"],
        "summary": "Trigger one daemon-owned PTR sync pass",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "responses": {
          "200": { "description": "Immediate PTR status after accepted trigger", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PTRStatusResponse" } } } },
          "400": { "description": "PTR disabled; status payload returned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PTRStatusResponse" } } } },
          "503": { "description": "PTR unavailable; status payload returned", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PTRStatusResponse" } } } },
          "501": { "description": "PTR trigger unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    },
    "/manage_database/integrity_check": {
      "post": {
        "tags": ["database"],
        "summary": "Run SQLite integrity check",
        "security": [{ "HydrusAccessKey": [] }, { "HydrusSessionKey": [] }],
        "responses": {
          "200": { "description": "Integrity check result", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/IntegrityResponse" } } } },
          "501": { "description": "Integrity checks unavailable", "content": { "text/plain": { "schema": { "type": "string" } } } }
        }
      }
    }
  }
}